requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
django-filter==2.4.0
django-psycopg2-extension==0.0.5
djangorestframework==3.12.4
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
